requirements.txt
# FastAPI Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
celery==5.3.6
redis==5.0.1
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
google-generativeai==0.3.2

# Streamlit Frontend Requirements
streamlit==1.30.0
pandas==2.1.4
plotly==5.18.0
extra-streamlit-components==0.1.70
